+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
   "packages/loadable-components",
   "packages/styled-components",
   "packages/styled-jsx",
+  "packages/plugin-styled-jsx",
   "packages/transform-imports",
   "packages/relay",
   "packages/noop",
