+++ packages/plugin-styled-jsx/Cargo.toml
@@ -1,10 +1,13 @@
 [package]
 authors = ["강동윤 <kdy1997.dev@gmail.com>"]
-description = "AST transforms visitor for styled-jsx"
+description = "SWC Plugin for styled-jsx"
 edition = "2021"
 license = "Apache-2.0"
-name = "styled_jsx"
+name = "swc_plugin_styled_jsx"
 version = "0.8.0"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
 
 [features]
 custom_transform = ["swc_core/common_concurrent"]
@@ -26,6 +29,8 @@ swc_core = { version = "0.38.1", features = [
   "ecma_utils",
   "ecma_visit"
 ] }
+
+styled_jsx = { version = "0.8.0", path = "../styled-jsx" }
 
 [dev-dependencies]
 testing = "0.31.8"
