+++ packages/plugin-styled-jsx/package.json
